Implimenting Authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AddQuestionValidationRequest;
 use App\Models\Question;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -62,6 +62,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403,"Access denied");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -74,6 +77,9 @@ class QuestionsController extends Controller
      */
     public function update(AddQuestionValidationRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403,"Access denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your Question has been updated !');
     }
@@ -86,6 +92,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403,"Access denied");
+        }
         $question->delete();
         return redirect('/questions')->with('success', 'Your Question has been removed');
     }
